% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_capitales_prov.R
\docType{data}
\name{capitales_prov}
\alias{capitales_prov}
\title{Capitales de las 52 "Provincias" de Spain}
\format{A data frame with 52 rows y 4 variables:
\itemize{
\item INECodMuni: Código municipio más usado (5 digitos)
\item NOMBRE_ACTUAL: Nombre del municipio
\item COD_PROV: Código de la provincia.
\item PROVINCIA: Nombre de la provincia.
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
capitales_prov
}
\description{
df con los 52 municipios Spain que son capitales de provincia (+ Ceuta y Melilla)
}
\examples{
\dontrun{
 df <- capitales_prov
}

}
\keyword{datasets}
