% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IGN.R
\docType{data}
\name{IGN_nomencla_condominios}
\alias{IGN_nomencla_condominios}
\title{Nomenclator del IGN de Condominios o COMUNIDADES JURISDICCIONLES}
\format{A data frame with 81 rows y 12 variables:
\itemize{
\item IdCon:Identificador del Comunidad Jurisdiccional.
\item NOMBRE_ACTUAL: Nombre de la Comunidad Jurisdiccional.
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
IGN_nomencla_condominios
}
\description{
Esta tabla contiene aquellos territorios con superficie propia que cuya gestión es llevada a
cabo por varios municipios, provincias o es directamente estatal.
}
\examples{
\dontrun{
 df <- IGN_nomencla_condominios
}

}
\keyword{datasets}
