% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IGN.R
\docType{data}
\name{IGN_nomencla_eatims}
\alias{IGN_nomencla_eatims}
\title{Nomenclator del IGN de EATIMS}
\format{A data frame with 3.708 rows y 13 variables:
\itemize{
\item CODINSCRIP: Número de inscripción de la EATIM en el Registro de Entidades Locales
\item CODPROV: Numeración correspondiente a la provincia en la que se encuentra la EATIM
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
IGN_nomencla_eatims
}
\description{
Esta tabla contiene la relación de Entidades de Ámbito Territorial Inferior al Municipio. La
fuente de datos de referencia en este caso es el Registro de Entidades Locales, completada con
información procedente del Instituto Nacional de Estadística y el Instituto Geográfico Nacional.
}
\examples{
\dontrun{
 df <- IGN_nomencla_eatims
}

}
\keyword{datasets}
