% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IGN.R
\docType{data}
\name{IGN_nomencla_entidades}
\alias{IGN_nomencla_entidades}
\title{Nomenclator del IGN de ENTIDADES}
\format{A data frame with 153.210 rows y 15 variables:
\itemize{
\item CODIGOINE: Código de 11 digitos (del INE, unico para cada entidad poblacional)
\item NOMBRE: Nombre de la unidad poblacional.
\item CODPROV: Código de la provincia.
\item NOMBRE_PROVINCIA: Nombre de la provincia.
\item TIPO: Clasificación interna de los tipos de entidad
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
IGN_nomencla_entidades
}
\description{
Tabla que contiene toda la información descriptiva de las entidades de población.
La fuente de referencia son los datos publicados por el Instituto Nacional de Estadística a través
de su nomenclátor poblacional completada con información procedente del Registro de
Entidades Locales y del Instituto Geográfico Nacional.
}
\examples{
\dontrun{
 df <- IGN_nomencla_entidades
}

}
\keyword{datasets}
