% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IGN.R
\docType{data}
\name{IGN_nomencla_muni}
\alias{IGN_nomencla_muni}
\title{Nomenclator del IGN de MUNICIPIOS}
\format{A data frame with 8.124 rows y 17 variables:
\itemize{
\item CODIGOINE: Código de 11 digitos (del INE, unico para cada entidad poblacional)
\item ID_REL: Identificador dado por el Registro de Entidades Locales. No todas las unidades poblacionales disponen de este identificador. Este campo esta en permanecen actualización.
\item CODPROV: Código de la provincia.
\item NOMBRE_PROVINCIA: Nombre de la provincia.
\item NOMBRE: Nombre actual del municipio tal y como aparece inscrito en el REL
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
IGN_nomencla_muni
}
\description{
Tabla que contiene toda la información descriptiva de los municipios. La fuente de datos de
referencia es el Registro de Entidades Locales, completada con información procedente del
Instituto Nacional de Estadística y el Instituto Geográfico Nacional.
Codigos y nombres de los municipios. tambien hay altitud, perimetro, superficie y algo mas
}
\examples{
\dontrun{
 df <- IGN_nomencla_muni
}

}
\keyword{datasets}
