% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IGN.R
\docType{data}
\name{IGN_nomencla_prov}
\alias{IGN_nomencla_prov}
\title{Nomenclator del IGN de PROVINCIAS}
\format{A data frame with 52 rows y 5 variables:
\itemize{
\item CODPROV: C칩digo de la provincia.
\item NOMBRE_PROVINCIA: Nombre de la provincia.
\item CAPITAL: Nombre de la capitalidad de la provincia.
}}
\source{
\url{http://www.ign.es/web/ign/portal/rcc-nomenclator-nacional}
}
\usage{
IGN_nomencla_prov
}
\description{
Tabla que contiene las denominaciones de las provincias espa침olas, junto a su c칩digo
provincial y la comunidad aut칩noma a la que pertenecen.
Tiene la capital de Provincia (good)
}
\examples{
\dontrun{
 df <- IGN_nomencla_prov
}

}
\keyword{datasets}
