% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanishRentidadesIGN-package.R
\docType{package}
\name{spanishRentidadesIGN-package}
\alias{spanishRentidadesIGN}
\alias{spanishRentidadesIGN-package}
\title{spanishRentidadesIGN: Facilita el uso del Nomneclator de entidades del IGN.}
\description{
El package spanishRentidadesIGNMore facilita el uso en R de los datos del
NOMENCLATOR DE MUNICIPIOS Y ENTIDADES DE POBLACIÓN del Instituto Geografico nacional (IGN).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perezp44/spanishRentidadesIGN}
  \item Report bugs at \url{https://github.com/perezp44/spanishRentidadesIGN/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
